agent-manager/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
kubernetes==28.1.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
celery==5.3.4
prometheus-client==0.19.0
google-cloud-logging==3.8.0
google-cloud-monitoring==2.16.0
google-cloud-bigquery==3.13.0
websockets==12.0
aiofiles==23.2.1
httpx==0.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
pyyaml==6.0.1
jinja2==3.1.2
docker==6.1.3
grpcio==1.59.3
protobuf==4.25.1